Move bedrock LLMObs tests to separate file
